+++ src/Workspaces/Core/Portable/Differencing/LongestCommonSubsequence.cs
@@ -108,7 +108,7 @@ namespace Microsoft.CodeAnalysis.Differencing
 
             int max = Math.Max(oldLength, newLength);
             Debug.Assert(lcsLength <= max);
-            return 1.0 - lcsLength / (double)max;
+            return 1.0 - (double)lcsLength / (double)max;
         }
 
         /// <summary>
